Made FILE_EXTENSION_PATTERN project-wide constant
