+++ tests/test_args_handling.py
@@ -9,43 +9,10 @@ import pytest
 
 # pylint: disable=C0116, W0611
 from auto_file_sorter.args_handling import (
-    _FILE_EXTENSION_PATTERN,
     _add_to_startup,
     handle_locations_args,
     resolved_path_from_str,
 )
-
-
-@pytest.mark.parametrize(
-    "extension",
-    (
-        pytest.param(".TXT"),
-        pytest.param(".zip"),
-        pytest.param(".7z"),
-        pytest.param(".123"),
-        pytest.param(".PnG"),
-        pytest.param(".Jp2"),
-    ),
-)
-def test_file_valid_extension_pattern(extension: str) -> None:
-    assert _FILE_EXTENSION_PATTERN.fullmatch(extension) is not None
-
-
-@pytest.mark.parametrize(
-    "extension",
-    (
-        pytest.param("TXT"),
-        pytest.param("zip"),
-        pytest.param("_7z_"),
-        pytest.param("-123"),
-        pytest.param(".PnG!"),
-        pytest.param(".Jp2@"),
-        pytest.param("/doc"),
-        pytest.param(""),
-    ),
-)
-def test_file_invalid_extension_pattern(extension: str) -> None:
-    assert _FILE_EXTENSION_PATTERN.fullmatch(extension) is None
 
 
 @pytest.mark.parametrize(
